fix: cast travel_broadcast_timeout_minutes to int
env() returns raw strings for set variables (only true/false/null/empty
are auto-cast), so a non-numeric TRAVEL_BROADCAST_TIMEOUT value flowed
straight into Carbon::addMinutes() in TravelService::process(), which
requires int|float and throws a TypeError otherwise — the root cause of
the 500 on POST /api/v2/travelanalyzer.

+++ config/pso-services.php
@@ -48,7 +48,7 @@ return [
         'do_in_locality_incentive' => 1,
         'timezone' => env('PSO_TIMEZONE', 'America/Toronto'),
         'travel_broadcast_api' => env('TRAVEL_BROADCAST_API', 'https://pso-services-6g6mj.kinsta.app/api/v2/travelanalyzerservice'),
-        'travel_broadcast_timeout_minutes' => env('TRAVEL_BROADCAST_TIMEOUT', 2),
+        'travel_broadcast_timeout_minutes' => (int) env('TRAVEL_BROADCAST_TIMEOUT', 2),
     ],
 
     /*
